+++ src/main.rs
@@ -13,7 +13,7 @@ fn ray_color(r: &ray::Ray, world: &dyn hittable::Hittable, depth: u32) -> color:
     }
 
     let mut rec = hittable::HitRecord::new();
-    if world.hit(&r, &0.001, &f32::INFINITY, &mut rec) {
+    if world.hit(&r, &0.001, &f64::INFINITY, &mut rec) {
         let mut scattered = ray::Ray::zero();
         let mut attenuation = color::Color::zero();
         if rec
@@ -48,8 +48,8 @@ fn render(
         for i in 0..*image_width {
             let mut pixel_color = color::Color::zero();
             for _s in 0..*samples_per_pixel {
-                let u = (i as f32 + utils::random_double(&0.0, &1.0)) / (image_width - 1) as f32;
-                let v = (j as f32 + utils::random_double(&0.0, &1.0)) / (image_height - 1) as f32;
+                let u = (i as f64 + utils::random_double(&0.0, &1.0)) / (image_width - 1) as f64;
+                let v = (j as f64 + utils::random_double(&0.0, &1.0)) / (image_height - 1) as f64;
                 let r = cam.get_ray(&u, &v);
                 pixel_color += &ray_color(&r, world, *max_depth);
             }
@@ -69,12 +69,12 @@ fn main() {
     // Image
     let aspect_ratio = 16.0 / 9.0;
     let image_width = 400 as usize;
-    let image_height = (image_width as f32 / aspect_ratio) as usize;
+    let image_height = (image_width as f64 / aspect_ratio) as usize;
     let samples_per_pixel = 100 as usize;
     let max_depth = 50 as u32;
 
     // World
-    let R = (std::f32::consts::PI / 4.0).cos();
+    let R = (std::f64::consts::PI / 4.0).cos();
     let mut world = hittable::HittableList::new_empty();
 
     let material_ground =
