+++ src/ray.rs
@@ -1,15 +1,15 @@
 use super::utils;
 use super::vec3;
 
-pub type Point = vec3::Vec3<f32>;
-pub type Vector = vec3::Vec3<f32>;
+pub type Point = vec3::Vec3<f64>;
+pub type Vector = vec3::Vec3<f64>;
 
-impl vec3::Vec3<f32> {
+impl vec3::Vec3<f64> {
     pub fn zero() -> Self {
         return Self::new(0.0, 0.0, 0.0);
     }
 
-    pub fn random(min: &f32, max: &f32) -> Self {
+    pub fn random(min: &f64, max: &f64) -> Self {
         return Self::new(
             utils::random_double(min, max),
             utils::random_double(min, max),
@@ -64,7 +64,7 @@ pub fn reflect(v: &Vector, n: &Vector) -> Vector {
     return v - n * 2.0 * vec3::dot(v, n);
 }
 
-pub fn refract(uv: &Vector, n: &Vector, etai_over_etat: &f32) -> Vector {
+pub fn refract(uv: &Vector, n: &Vector, etai_over_etat: &f64) -> Vector {
     let cos_theta = vec3::dot(&-uv, n).min(1.0);
     let r_out_perp = (uv + n * cos_theta) * etai_over_etat;
     let r_out_parallel = n * -(1.0 - r_out_perp.length_squared()).abs().sqrt();
@@ -96,7 +96,7 @@ impl Ray {
         return &self.dir;
     }
 
-    pub fn at(&self, t: &f32) -> Point {
+    pub fn at(&self, t: &f64) -> Point {
         return self.origin() + self.direction() * t.clone();
     }
 }
