+++ src/utils.rs
@@ -1,6 +1,6 @@
 use rand::Rng;
 
-pub fn random_double(min: &f32, max: &f32) -> f32 {
+pub fn random_double(min: &f64, max: &f64) -> f64 {
     thread_local!(static RNG_STORAGE : std::cell::RefCell<rand::prelude::ThreadRng> 
         = std::cell::RefCell::new(rand::thread_rng()));
     return RNG_STORAGE.with(|rng_ref| return rng_ref.borrow_mut().gen_range(*min..*max));
